Add a helper type for finding multi-step reference frame transformations (#111)
We've discussed adding this, and now I've had a real need to do this.
This is the MVP implementation for helping find the "shortest path"
series of transformations. This is accomplished by building up a graph
with the reference frames as vertices and the transformations as edges,
then performing a breadth-first search through this graph for the first
available path.

# Long Term Improvements
I've implemented this functionality as an additional function to avoid
breaking API changes. In a future revision I would like to roll up this
functionality into the `get_transformation()` function and make it
transparent, but doing that in the way I am thinking would require
adding a new type representing a compound transformation and modifying
the `Transformation` struct to be able to use it.