Add LaTeX formatting to doc comments (#118)
There are a few places where a bit of LaTeX would help make things a bit
more clear in our documentation. Indeed the C library documentation has
several uses of LaTeX which Doxygen handles. This is based on the
[rustdoc-katex-demo](https://github.com/paulkernfeld/rustdoc-katex-demo)
crate, which demonstates how to include a link to the KaTex code in the
HTML generated for the documents.