Remove other functionality not needed in Rust (#129)
The final chunk of changes from #122 🎉 

Simply put this removes the remaining functionality that depends on the
C library and breaks the dependency on the `swiftnav-sys` crate. I
wouldn't say that we'll never want this functionality in Rust, but these
modules have not been meaningfully used up till now and have non-trivial
C code backing them so would require an undue amount of work to
re-write.

# What's been changed
- Crate level documentation has been updated to not reference removed
functionality, and to give a description of functionality originally
missed
- Minimum supported Rust version (MSRV) has been bumped up to reflect
reality (also fix CI build to check MSRV still builds)

# What's been removed
- `ephemeris` module - Decoding and calculating Satellite positions
- `geoid` module - Calculating difference between Ellipsoid and Geoid
- `ionosphere` module - Decoding and evaluating Klobuchar ionosphere
model
- `troposphere` module - Calculating troposphere delays
- `navmeas` module - Datatypes related to storing GNSS measurments
- `solver` module - Single epoch position solver
- All conversions to/from `swiftnav_sys` types, all of these were
private

# What's been added
- N/A