Set up pedantic Clippy warnings (#131)
This sets the `clippy::pedantic` set of lints by default to be warnings,
but remember that CI will fail on any warnings. The second commit goes
through and fixes most of the new warnings. There are a handful of
places where some of the lints are disabled.

## Disabled lints

- `#[allow(clippy::many_single_char_names)]` was used twice
(`crate::coords::llh::LLHRadians::to_ecef()`, and
`crate::math::compile_time_sqrt()`). These two functions are rather
short and implement common mathematical formulas, so there aren't really
super descriptive names to give the variables and doing so would make
the implementation even more different than the original source.
- `#![allow(clippy::cast_possible_truncation, clippy::cast_sign_loss,
clippy::cast_precision_loss)]` for the entire `crate::time::mjd` module,
basically we're doing a lot of not great casting
- `#![allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]`
for the entire `crate::time::utc` module, similar use of lots of not
great casting