Add nice to have functions for coordindate types (#130)
Think I got a little over zealous with the RIIR, and removed a couple of
nice to have functions.

Is it acceptable to have multiple `AsRef` impls on a type? It seems like
the stdlin does (e.g. `Vec` implements two different `AsRef`). Also is
it preferable to not have dedicated functions and only rely on the
`AsRef` impl for casting?